handle most of the keys
