+++ examples/bevy.rs
@@ -8,6 +8,7 @@ use bevy::{
     log::{info, LogPlugin},
 };
 use dioxus::prelude::*;
+use std::convert::From;
 
 #[derive(Debug, Clone)]
 enum CoreCommand {
@@ -16,16 +17,33 @@ enum CoreCommand {
 }
 
 #[derive(Clone, Copy, Debug, PartialEq)]
+// mimic bevy's KeyboardInput since it doesn't implement Copy.
 pub struct KeyboardInput {
     pub scan_code: u32,
     pub key_code: Option<KeyCode>,
     pub state: ElementState,
 }
 
+impl From<&BevyKeyboardInput> for KeyboardInput {
+    fn from(input: &BevyKeyboardInput) -> Self {
+        KeyboardInput {
+            scan_code: input.scan_code,
+            key_code: input.key_code,
+            state: input.state,
+        }
+    }
+}
+
 #[derive(Debug, Clone, Copy)]
 enum UICommand {
     Test,
     KeyboardInput(KeyboardInput),
+}
+
+impl From<&BevyKeyboardInput> for UICommand {
+    fn from(input: &BevyKeyboardInput) -> Self {
+        UICommand::KeyboardInput(input.into())
+    }
 }
 
 fn main() {
@@ -36,23 +54,22 @@ fn main() {
         .add_plugin(DioxusDesktopPlugin::<CoreCommand, UICommand>::new(app, ()))
         .insert_non_send_resource(config)
         .add_plugin(LogPlugin)
-        .add_startup_system(setup)
         .add_system(handle_core_command)
         .add_system(send_keyboard_input)
         .run();
 }
 
-fn setup(mut event: EventWriter<UICommand>) {
-    event.send(UICommand::Test);
-}
-
-fn handle_core_command(mut events: EventReader<CoreCommand>, mut event: EventWriter<AppExit>) {
+fn handle_core_command(
+    mut events: EventReader<CoreCommand>,
+    mut exit: EventWriter<AppExit>,
+    mut ui: EventWriter<UICommand>,
+) {
     for cmd in events.iter() {
         info!("🧠 {:?}", cmd);
 
         match cmd {
-            CoreCommand::Quit => event.send(AppExit),
-            _ => {}
+            CoreCommand::Quit => exit.send(AppExit),
+            CoreCommand::Test => ui.send(UICommand::Test),
         }
     }
 }
@@ -64,26 +81,24 @@ fn send_keyboard_input(
     for input in events.iter() {
         info!("🧠 {:?}", input);
 
-        // copy KeyboardInput and send to UI
-        event.send(UICommand::KeyboardInput(KeyboardInput {
-            scan_code: input.scan_code,
-            key_code: input.key_code,
-            state: input.state,
-        }));
+        event.send(input.into());
     }
 }
 
 fn app(cx: Scope) -> Element {
     let window = use_bevy_window::<CoreCommand, UICommand>(&cx);
+    let ui_cmd = use_state(&cx, || None);
     let keyboard_input = use_state(&cx, || None);
 
     use_future(&cx, (), |_| {
         let keyboard_input = keyboard_input.clone();
+        let ui_cmd = ui_cmd.clone();
         let mut rx = window.receiver();
 
         async move {
             while let Ok(cmd) = rx.recv().await {
-                info!("🎨 {:?}", cmd);
+                *ui_cmd.make_mut() = Some(cmd);
+
                 match cmd {
                     UICommand::KeyboardInput(input) => {
                         *keyboard_input.make_mut() = Some(input);
@@ -94,7 +109,7 @@ fn app(cx: Scope) -> Element {
         }
     });
 
-    cx.render(rsx! {
+    cx.render(rsx! (
         div {
             div {
                 h1 { "Bevy Dioxus Plugin Example" },
@@ -115,17 +130,24 @@ fn app(cx: Scope) -> Element {
                     "Minimize"
                 }
             }
-            keyboard_input.and_then(|input| {
-                let input = format!("{:#?}", input);
-                Some(rsx! {
-                    div {
-                        h2 { "Keyboard Input" },
-                        code {
-                            "{input}"
-                        }
-                    }
-                })
-            })
-        }
-    })
+
+            ui_cmd.and_then(|cmd| Some(rsx!(
+                div {
+                     h2 { "UI Command" },
+                     div {
+                         box_sizing: "border-box",
+                         background: "#DCDCDC",
+                         height: "4rem",
+                         width: "100%",
+                         display: "flex",
+                         align_items: "center",
+                         border_radius: "4px",
+                         padding: "1rem",
+                         code {
+                             [format_args!("{:#?}", cmd)],
+                         }
+                     }
+                }
+            )))
+    }))
 }
