+++ packages/bevy/src/event.rs
@@ -5,6 +5,12 @@ use bevy::input::{
 use serde::Deserialize;
 use serde_json::Value;
 
+#[derive(Debug, Clone)]
+pub enum CustomUserEvent<CoreCommand> {
+    CoreCommand(CoreCommand),
+    KeyboardInput(KeyboardInput),
+}
+
 #[derive(Debug, Deserialize)]
 #[serde(tag = "type")]
 enum WebKeyboardEvent {
@@ -130,22 +136,21 @@ pub fn try_parse_code(code: String) -> Option<KeyCode> {
         "Compose" => Some(KeyCode::Compose),
 
         // Caret,
-
-        // Numlock,
-        // Numpad0,
-        // Numpad1,
-        // Numpad2,
-        // Numpad3,
-        // Numpad4,
-        // Numpad5,
-        // Numpad6,
-        // Numpad7,
-        // Numpad8,
-        // Numpad9,
+        "NumLock" => Some(KeyCode::Numlock),
+        "Numpad0" => Some(KeyCode::Numpad0),
+        "Numpad1" => Some(KeyCode::Numpad1),
+        "Numpad2" => Some(KeyCode::Numpad2),
+        "Numpad3" => Some(KeyCode::Numpad3),
+        "Numpad4" => Some(KeyCode::Numpad4),
+        "Numpad5" => Some(KeyCode::Numpad5),
+        "Numpad6" => Some(KeyCode::Numpad6),
+        "Numpad7" => Some(KeyCode::Numpad7),
+        "Numpad8" => Some(KeyCode::Numpad8),
+        "Numpad9" => Some(KeyCode::Numpad9),
 
         // AbntC1,
         // AbntC2,
-        // NumpadAdd,
+        "NumpadAdd" => Some(KeyCode::NumpadAdd),
         "Quote" => Some(KeyCode::Apostrophe),
         // Apps,
         // Asterisk,
@@ -156,54 +161,50 @@ pub fn try_parse_code(code: String) -> Option<KeyCode> {
         // Calculator,
         // Capital,
         // Colon,
-        // Comma,
-        // Convert,
-        // NumpadDecimal,
-        // NumpadDivide,
-        // Equals,
-        // Grave,
+        "Comma" => Some(KeyCode::Comma),
+        "Convert" => Some(KeyCode::Convert),
+        "NumpadDecimal" => Some(KeyCode::NumpadDecimal),
+        "NumpadDivide" => Some(KeyCode::NumpadDivide),
+        "Equal" => Some(KeyCode::Equals),
+        "Backquote" => Some(KeyCode::Grave),
         // Kana,
         // Kanji,
-        // /// The left alt key. Maps to left option on Mac.
-        // LAlt,
-        // LBracket,
-        // LControl,
-        // LShift,
-        // /// The left Windows key. Maps to left Command on Mac.
-        // LWin,
+        "AltLeft" => Some(KeyCode::LAlt),
+        "BracketLeft" => Some(KeyCode::LBracket),
+        "ControlLeft" => Some(KeyCode::LControl),
+        "ShiftLeft" => Some(KeyCode::LShift),
+        "MetaLeft" => Some(KeyCode::LWin),
         // Mail,
         // MediaSelect,
         // MediaStop,
-        // Minus,
-        // NumpadMultiply,
+        "Minus" => Some(KeyCode::Minus),
+        "NumpadMultiply" => Some(KeyCode::NumpadMultiply),
         // Mute,
         // MyComputer,
-        // NavigateForward,  // also called "Prior"
-        // NavigateBackward, // also called "Next"
+        // "BrowserForward" => Some(KeyCode::NavigateForward),
+        // "BrowserBackward" => Some(KeyCode::NavigateBackward),
         // NextTrack,
-        // NoConvert,
-        // NumpadComma,
-        // NumpadEnter,
-        // NumpadEquals,
+        "NonConvert" => Some(KeyCode::NoConvert),
+        "NumpadComma" => Some(KeyCode::NumpadComma),
+        "NumpadEnter" => Some(KeyCode::NumpadEnter),
+        "NumpadEqual" => Some(KeyCode::NumpadEquals),
         // Oem102,
-        // Period,
+        "Period" => Some(KeyCode::Period),
         // PlayPause,
-        // Power,
+        "Power" => Some(KeyCode::Power),
         // PrevTrack,
-        // /// The right alt key. Maps to right option on Mac.
-        // RAlt,
-        // RBracket,
-        // RControl,
-        // RShift,
-        // /// The right Windows key. Maps to right Command on Mac.
-        // RWin,
-        // Semicolon,
-        // Slash,
+        "AltRight" => Some(KeyCode::RAlt),
+        "BracketRight" => Some(KeyCode::RBracket),
+        "ControlRight" => Some(KeyCode::RControl),
+        "ShiftRight" => Some(KeyCode::RShift),
+        "MetaRight" => Some(KeyCode::RWin),
+        "Semicolon" => Some(KeyCode::Semicolon),
+        "Slash" => Some(KeyCode::Slash),
         // Sleep,
         // Stop,
-        // NumpadSubtract,
+        "NumpadSubtract" => Some(KeyCode::NumpadSubtract),
         // Sysrq,
-        // Tab,
+        "Tab" => Some(KeyCode::Tab),
         // Underline,
         // Unlabeled,
         // VolumeDown,
@@ -216,10 +217,11 @@ pub fn try_parse_code(code: String) -> Option<KeyCode> {
         // WebRefresh,
         // WebSearch,
         // WebStop,
-        // Yen,
-        // Copy,
-        // Paste,
-        // Cut,
+        "IntlYen" => Some(KeyCode::Yen),
+        "Copy" => Some(KeyCode::Copy),
+        "Paste" => Some(KeyCode::Paste),
+        "Cut" => Some(KeyCode::Cut),
+
         _ => None,
     }
 }
